app/requirements.txt
gradio==3.41.2
google-cloud-aiplatform==1.31.1
google-cloud-logging==3.6.0